% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_downloads.R
\name{download_log}
\alias{download_log}
\title{Download and unzip a package download log .csv}
\usage{
download_log(day, package_name = NULL, cache_dir = NULL)
}
\arguments{
\item{day}{day of the log to download}

\item{package_name}{package name if not NULL, downloads are directly filtered for this package}

\item{cache_dir}{if not NULL, results are cached in this directory}
}
\description{
Format is assumed to be as provided by cran-logs.rstudio.com
}

