% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_downloads.R
\name{package_download_cache_dir}
\alias{package_download_cache_dir}
\alias{remove_package_download_cache_dir}
\title{Canonical download directory for package download logs}
\usage{
package_download_cache_dir(package_name)

remove_package_download_cache_dir(package_name)
}
\arguments{
\item{package_name}{name of package}
}
\description{
Canonical download directory for package download logs
}

